requirements.txt
openai-agents
requests
fastmcp
mcp
httpx
python-dotenv